reveiew comments

+++ apps/frobots/lib/frobots/assets/cannon.ex
@@ -18,7 +18,7 @@ defmodule Frobots.Assets.Cannon do
   schema "cannons" do
     field :type, Ecto.Enum, values: ~w(Mk1 Mk2 Mk3)a
     field :reload_time, :integer
-    field :rate_of_fire, :float
+    field :rate_of_fire, :integer
     field :magazine_size, :integer
     field :image, :string, default: "https://via.placeholder.com/50.png"
     has_many :cannon_inst, Frobots.Assets.CannonInst
